requirements.txt
groq==0.11.0
langchain==0.3.3
langchain-community==0.3.2
langchain-groq==0.2.0
python-dotenv==1.0.1
